requirements.txt
flask
rembg==2.0.57
onnxruntime==1.12.0
Pillow
pytest